fix: break cyclic import between action_devices and notifications
Agent-Logs-Url: https://github.com/DaTiC0/smart-google/sessions/a7c83973-d307-4a35-8591-9afa75f5f1b1
