+++ action_devices.py
@@ -7,6 +7,11 @@ import requests
 import secrets
 from flask import current_app
 from notifications import mqtt
+from firebase_utils import (
+    _normalize_user_scope,
+    _get_user_device_states_ref,
+    reference,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -17,84 +22,6 @@ except ImportError:
     state = None
     REPORTSTATE_AVAILABLE = False
     logger.warning("ReportState module not available, some features may be disabled.")
-
-# Try to import firebase_admin, but provide fallback if not available
-try:
-    from firebase_admin import db
-    FIREBASE_AVAILABLE = True
-except ImportError:
-    FIREBASE_AVAILABLE = False
-    logger.warning("Firebase admin not available, using mock data for testing")
-
-# Mock data for testing when Firebase is not available
-MOCK_DEVICES = {
-    "test-light-1": {
-        "type": "action.devices.types.LIGHT",
-        "traits": ["action.devices.traits.OnOff", "action.devices.traits.Brightness"],
-        "name": {"name": "Test Light"},
-        "willReportState": True,
-        "attributes": {"colorModel": "rgb"},
-        "states": {"on": True, "brightness": 80, "online": True}
-    },
-    "test-switch-1": {
-        "type": "action.devices.types.SWITCH",
-        "traits": ["action.devices.traits.OnOff"],
-        "name": {"name": "Test Switch"},
-        "willReportState": True,
-        "states": {"on": False, "online": True}
-    }
-}
-
-
-class MockRef:
-    @staticmethod
-    def get():
-        return MOCK_DEVICES
-
-    @staticmethod
-    def child(path):
-        return MockChild(MOCK_DEVICES, path)
-
-
-class MockChild:
-    def __init__(self, data, path):
-        self.data = data
-        self.path = path
-
-    def child(self, child_path):
-        return MockChild(self.data, self.path + '/' + child_path)
-
-    def get(self):
-        keys = self.path.split('/')
-        current = self.data
-        for key in keys:
-            if isinstance(current, dict) and key in current:
-                current = current[key]
-            else:
-                return None
-        return current
-
-    def update(self, values):
-        keys = self.path.split('/')
-        current = self.data
-        for key in keys[:-1]:
-            if key not in current:
-                current[key] = {}
-            current = current[key]
-        if keys[-1] not in current:
-            current[keys[-1]] = {}
-        current[keys[-1]].update(values)
-        return current[keys[-1]]
-
-
-def _normalize_user_scope(user_id):
-    """Normalize user scope value used in Firebase path building."""
-    if user_id is None:
-        return None
-    user_value = str(user_id).strip()
-    if not user_value or '/' in user_value or '\\' in user_value or '..' in user_value:
-        return None
-    return user_value
 
 
 def _to_sync_device(device_id, raw_data):
@@ -122,34 +49,12 @@ def _build_sync_devices(snapshot):
     return devices
 
 
-# firebase initialisation problem was fixed?
-def reference(user_id=None):
-    if FIREBASE_AVAILABLE:
-        try:
-            user_scope = _normalize_user_scope(user_id)
-            if user_scope:
-                return db.reference(f'/users/{user_scope}/devices')
-            return db.reference('/devices')
-        except Exception as e:
-            # Firebase is installed but not initialized (e.g. missing credentials in dev)
-            logger.warning("Firebase not initialized, falling back to mock data: %s", e)
-    return MockRef()
-
-
 def _get_scoped_snapshot(user_id):
     """Get per-user device snapshot from scoped Firebase path."""
     user_scope = _normalize_user_scope(user_id)
     if not user_scope:
         return {}
     return reference(user_scope).get() or {}
-
-
-def _get_user_device_states_ref(user_id, device_id):
-    """Centralized helper for building user-scoped device state Firebase references."""
-    user_scope = _normalize_user_scope(user_id)
-    if not user_scope or not device_id or '/' in str(device_id):
-        return None
-    return reference(user_scope).child(str(device_id)).child('states')
 
 
 def _get_scoped_device_states(device_id, user_id):
